Add files via upload
As a first-year university student, our team's initial idea was to save time for people while they wait at traffic lights. In this project, we initially decided to implement it using electronic and Arduino components. However, due to limited resources, we were unable to proceed with this approach. Therefore, we decided to create a manual Java form for generating numbers instead.